Change version detection to permit >= current
Permit versions greater than or equal to the "known" version.

For both iOS and Android, the version_id parameter is supplied as an
integer-like value from the client, and is intended to increase with
each new release. If the client submits a version that is less than
the version specified here, we provide an update message.

If the version_id is equal to or greater than the version specified
here, no update message is returned.

This allows us to have a brand new build show "updates available",
and permits us to have some overlap where the latest version may be
available from distribution sources, but current users aren't
notified until we are certain that the application is available to
all users.

Signed-off-by: Jeff Godin <jgodin@tadl.org>